App successfully works painting th picturte based on user top song meta data. To further the project, implement a way where the user's data is immediately ready to go from launching the application instead of waiting. I have some ideas about how to do this such as implementing a splash screen to retrieve the data at the start
Last, I need to make the app look somewhat presentable. I need to use
some software to create a nice user interface. That is not a priority
yet however.

Overall, this commit seems pretty hopeful and stable.

Latewr down the road can the integration with an LED strip come into
the environment since there is meta data on each track (audio analysis)
that may be helpful in lighting up a strip. (This is a long message)