Added a safe way (I think) to ensure the track cover art and audio features update whenever a new song plays. I still need to update this when the user is in the UserStatistcs activity.
Fixed the error where the wrong stats would show if a user clicked to
fast on the next activity and also the stuck activity where it would
just display loading on the UI for the track audio features.

Lastly, I was able to get the current playing artist profile picture to
show up on the homescreen whenever the song changes and a new artist is
present.

Future improvement: Make all the above things mentioned faster.